app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.5.0",
    "msgspec>=0.19.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]
